chore: refactor kube pvc columns
Refactors the Kubernetes PVC columns using /columns and a props.ts.

Part of #11035.
