--- packages/renderer/src/lib/pvc/PVCColumnActions.svelte
@@ -1,11 +1,0 @@
-<script lang="ts">
-import PVCActions from './PVCActions.svelte';
-import type { PVCUI } from './PVCUI';
-
-interface Props {
-  object: PVCUI;
-}
-let { object }: Props = $props();
-</script>
-
-<PVCActions pvc={object} />
