+++ packages/renderer/src/lib/pvc/columns/Actions.svelte
@@ -1,0 +1,8 @@
+<script lang="ts">
+import PVCActions from '../PVCActions.svelte';
+import type { Props } from './props';
+
+let { object }: Props = $props();
+</script>
+
+<PVCActions pvc={object} />
